data/requirements.txt
# Database requirements for Strategic Intelligence App
# PostgreSQL database integration

# Core SQLAlchemy and PostgreSQL driver
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Additional database utilities
alembic>=1.12.0  # For database migrations
sqlalchemy-utils>=0.41.0  # Utility functions for SQLAlchemy

# For JSON serialization and datetime handling
python-dateutil>=2.8.0

# Development and testing
pytest>=7.0.0  # For testing database operations
pytest-asyncio>=0.21.0  # For async testing if needed

# Logging and monitoring
structlog>=23.0.0  # Enhanced logging capabilities

# Note: PgBouncer is a separate PostgreSQL application, not a Python package
# If you want connection pooling, install PgBouncer separately on your system 